test(snapshot): split openclaw config restore coverage
